+++ src/backend/user-server/.gitignore
@@ -35,3 +35,6 @@ out/
 
 ### VS Code ###
 .vscode/
+
+### DotEnv ###
+.env
