+++ src/backend/user-server/src/main/java/com/bbebig/userserver/friend/dto/response/FriendResponse.java
@@ -1,7 +1,7 @@
-package com.bbebig.user_server.presentation.dto;
+package com.bbebig.userserver.friend.dto.response;
 
-import com.bbebig.user_server.domain.Friend;
-import com.bbebig.user_server.domain.Member;
+import com.bbebig.userserver.friend.entity.Friend;
+import com.bbebig.userserver.member.entity.Member;
 
 import java.time.LocalDateTime;
