+++ src/backend/user-server/src/main/java/com/bbebig/userserver/friend/entity/FriendStatus.java
@@ -1,4 +1,4 @@
-package com.bbebig.user_server.domain;
+package com.bbebig.userserver.friend.entity;
 
 public enum FriendStatus {
     PENDING,DECLINED,ACCEPTED,DELETED
