+++ src/backend/user-server/src/main/java/com/bbebig/userserver/member/dto/request/MemberModifyRequest.java
@@ -1,4 +1,4 @@
-package com.bbebig.user_server.presentation.dto;
+package com.bbebig.userserver.member.dto.request;
 
 public record MemberModifyRequest(
         String profileImgUrl,
