+++ src/backend/user-server/src/main/java/com/bbebig/userserver/member/entity/CustomPresenceStatus.java
@@ -1,4 +1,4 @@
-package com.bbebig.user_server.domain;
+package com.bbebig.userserver.member.entity;
 
 public enum CustomPresenceStatus {
     ONLINE, OFFLINE, NOT_DISTURB, INVISIBLE
