+++ src/backend/user-server/src/main/java/com/bbebig/userserver/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.bbebig.user_server.domain;
+package com.bbebig.userserver.member.repository;
 
+import com.bbebig.userserver.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member,Long> {
