+++ src/backend/user-server/src/test/java/com/bbebig/userserver/MemberServerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.bbebig.user_server;
+package com.bbebig.userserver;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
